fix: refactor path utility functions to enhance security against path traversal attacks
