+++ apps/app/src/server/service/growi-bridge/index.ts
@@ -7,7 +7,7 @@ import unzipStream, { type Entry } from 'unzip-stream';
 import type Crowi from '~/server/crowi';
 import loggerFactory from '~/utils/logger';
 
-import { assertFileNameSafeForBaseDir } from '../../util/path-utils';
+import { assertFileNameSafeForBaseDir } from '../../util/safe-path-utils';
 import type { ZipFileStat } from '../interfaces/export';
 import { tapStreamDataByPromise } from './unzip-stream-utils';
