+++ apps/app/src/server/util/safe-path-utils.ts
